feat: add first contract
This gets tooling mostly in place for the first contract, and
aspirationally names the directory `contracts` so that more can be added
later.

The contract at this point is just a copy of the one from
create-near-app. But the point here is to make sure that `yarn start`
works correctly (it does, for one contract) and to make the sign in on
the home page work (it does, but the page doesn't change afterward).